# backend/docs/swagger.yaml
basePath: /
definitions:
  auth.LoginRequest:
    properties:
      email:
        example: joao@example.com
        type: string
      password:
        example: senha123
        minLength: 6
        type: string
    required:
    - email
    - password
    type: object
  auth.RegisterRequest:
    properties:
      email:
        example: joao@example.com
        type: string
      name:
        example: João Silva
        type: string
      password:
        example: senha123
        minLength: 6
        type: string
    required:
    - email
    - name
    - password
    type: object
  domain.AssetType:
    enum:
    - CRYPTO
    - STOCK
    - ETF
    type: string
    x-enum-varnames:
    - Crypto
    - Stock
    - ETF
  domain.OperationType:
    enum:
    - BUY
    - SELL
    type: string
    x-enum-varnames:
    - Buy
    - Sell
  domain.Transaction:
    properties:
      asset_symbol:
        description: Símbolo do ativo
        type: string
      asset_type:
        allOf:
        - $ref: '#/definitions/domain.AssetType'
        description: Tipo de ativo
      created_at:
        description: Data de criação
        type: string
      id:
        description: UUID
        type: string
      operation_date:
        description: Data da operação
        type: string
      price_in_brl:
        description: Preço em BRL
        type: number
      price_in_usd:
        description: Preço em USD
        type: number
      quantity:
        description: Quantidade
        type: number
      type:
        allOf:
        - $ref: '#/definitions/domain.OperationType'
        description: Tipo de operação
      usd_brl_rate:
        description: Taxa de câmbio USD/BRL
        type: number
      user_id:
        description: UUID do usuário
        type: string
    type: object
  domain.User:
    properties:
      created_at:
        type: string
      email:
        type: string
      id:
        type: string
      name:
        type: string
    type: object
  transaction.AddTransactionRequest:
    properties:
      asset_symbol:
        description: Símbolo do ativo
        example: AAPL
        type: string
      asset_type:
        allOf:
        - $ref: '#/definitions/domain.AssetType'
        description: Tipo de ativo
        enum:
        - CRYPTO
        - STOCK
        - ETF
      operation_date:
        description: Data da operação
        type: string
      price_in_usd:
        description: Preço em USD
        minimum: 0
        type: number
      quantity:
        description: Quantidade
        minimum: 0
        type: number
      type:
        allOf:
        - $ref: '#/definitions/domain.OperationType'
        description: Tipo de operação
        enum:
        - BUY
        - SELL
    required:
    - asset_symbol
    - asset_type
    - operation_date
    - price_in_usd
    - quantity
    - type
    type: object
host: localhost:8080
info:
  contact: {}
  description: API para o sistema IRGlobal
  title: IRGlobal API
  version: "1.0"
paths:
  /auth/login:
    post:
      consumes:
      - application/json
      description: Autentica o usuário no sistema
      parameters:
      - description: Dados do usuário
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/auth.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Autentica o usuário
      tags:
      - auth
  /auth/register:
    post:
      consumes:
      - application/json
      description: Registra um novo usuário no sistema
      parameters:
      - description: Dados do usuário
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/auth.RegisterRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/domain.User'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Registrar um novo usuário
      tags:
      - auth
  /transaction/add:
    post:
      consumes:
      - application/json
      description: Adiciona uma nova Transação ao sistema
      parameters:
      - description: Token de autenticação
        in: header
        name: Authorization
        required: true
        type: string
      - description: Dados da transação
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/transaction.AddTransactionRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Transação criada com sucesso
          schema:
            $ref: '#/definitions/domain.Transaction'
        "400":
          description: Dados inválidos
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Erro interno do servidor
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Adiciona uma nova Transação
      tags:
      - transaction
swagger: "2.0"